refactor: dto update price type

+++ pkg/inventory/routes/dto/inventory_dto.go
@@ -1,14 +1,14 @@
 package dto
 
 type CreateItemRequestBody struct {
-	Name     string `json:"name"`
-	Quantity int64  `json:"quantity"`
-	Price    int64  `json:"price"`
+	Name     string  `json:"name"`
+	Quantity int64   `json:"quantity"`
+	Price    float64 `json:"price"`
 }
 
 type UpdateItemRequestBody struct {
-	Id       int64  `json:"id"`
-	Name     string `json:"name"`
-	Quantity int64  `json:"quantity"`
-	Price    int64  `json:"price"`
+	Id       int64   `json:"id"`
+	Name     string  `json:"name"`
+	Quantity int64   `json:"quantity"`
+	Price    float64 `json:"price"`
 }
